requirements.txt
streamlit
pandas
requests
plotly
google-generativeai
tqdm
python-dotenv
tabulate
pytest
pytest-mock
